move handler functionality to router for handling event and subrouter for handling calls
